Add support for cell barcodes which are longer than 16 nucleotides.
Add support for cell barcodes which are longer than 16 nucleotides.
The old code overwrote the first nucleotides of the cell barcodes
with As for position 1 till (length cell barcode - 16) as cell
barcodes where converted with "convertNuclStrToInt32" and converted
back to "convertNuclInt32toString" and this would cause an overflow.

The number of barcodes is still limited to 2^32 after this patch.
UMIs are still limited to 16 nucleotides. Longer UMIs could be
supported in a similar way.

+++ source/ParametersSolo.h
@@ -13,7 +13,7 @@ public:
     uint32 umiS,umiL; //umi start,length
     uint32 bL; //total barcode length
     string soloCBwhitelist;
-    std::vector <uint32> cbWL;
+    std::vector <uint64> cbWL;
     string strandStr;
     int32 strand;
     //features
